chore: update aws service spec in a separate job (#26531)
Currently this is just part of the `Yarn upgrade` workflow. 
We want a separate workflow so the updates are visible in the CHANGELOG.

This should also reduce potential failures in the `Yarn upgrade` workflow and vice-versa, so they don't block each other.

Tested o fork: https://github.com/mrgrain/aws-cdk/actions/runs/5680204361

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*