chore(cdk-lib): cloud-assembly-schema tests are flaky, becaue they are not free of side-effects (#26568)
The attempted fix in #26551 did not work.

This is another attempt with a different approach.
Based on the fact that the failure occurs in `manifest.ts`, there appear to be circumstances for this file to be load in a context that is not fully initialized. Thus the `Cannot find name 'require'` error.
However runtime code like `require` is not needed to generate the schemas.
In fact the `manifest.ts` file isn't needed at all for this.

This change ensures that only the required files are loaded when generating each schema.
Hopefully this will bypass the erroneous code.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*