chore(lambda): update default node runtime for all partitions (#26475)
Node18 is the default runtime environment for lambdas in the [default](https://aws.amazon.com/it/blogs/compute/node-js-18-x-runtime-now-available-in-aws-lambda/), [China](https://www.amazonaws.cn/en/new/2023/amazon-lambda-adds-support-for-node-js-18-in-amazon-web-services-china-regions/), and [GovCloud](https://aws.amazon.com/it/about-aws/whats-new/2023/04/aws-lambda-node-js-18-govcloud-regions/) regions.

This fix updates the runtime default versions for constructs that were still using older versions, in particular:

- `LogRetention`
- `AwsCustomResource`
- `CustomResourceProvider`

Also, it updates the `DefaultCrNodeVersionMap` in the fact tables to reflect the updated values.

Closes #26461.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*