fix: broken cross-region reference in aws-route53
Fixes the cross region zone delegation custom resource in route53 by
reverting to using the v2 aws-sdk. This requires reverting the runtime
to use Node16 until we figure out the correct usage of the v3 sdk.

At this point I suspect it is an issue with the SDK but I have been
unable to reliably reproduce the bug, regardless of the parent or child
stack being in an opt-in region and vice versa

Fixes: #26593