fix(custom-resources): incorrect sdkv3 fallback
Fixes logic within the aws-sdk-v3-handler for AwsCustomResource to
behave more like the v2 handler, in so much as if `installLatestSdk`
fails, but it can still require the package it was attempting to
install, it will continue.

This makes the runtime more friendly to deploying your constructs across
regions that may or may not have access to npm within the lambda
runtime.