fix(triggers): executed on update even when executeOnHandlerChange is false (#26676)
Currently, trigger is re-executed every time when the handler function code or configuration changes even if executeOnHandlerChange is set to false. This is because executeOnHandlerChange prop is never used in the current implementation. 

This PR solves the issue by referring executeOnHandlerChange prop and skipping function invocation when executeOnHandlerChange is set to false.

Closes #25939

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*