fix(sam): CfnFunction events are not rendered (#26679)
Because of a mistake introduced into the SAM schema, the `AlexaSkill` event type doesn't have any required properties anymore.

When the `CfnFunction` is trying all the different event types in the type union that it supports, it will go through every type in alphabetical order and pick the first type that doesn't fail its validation.

After the schema change, the first type (`Alexa` which starts with an `A`) would therefore accept all types: no required fields, and for JavaScript compatibility purposes we allow superfluous fields, and so we pick a type that doesn't render anything.

This change reorders the alternatives in the union such that stronger types are tried first.

`HttpApiEvent` and `AlexaSkillEvent` both have no required properties, and this now reverses the problem: `AlexaSkillEvent` can no longer be specified because `HttpApiEvent` will match first.

But that's the more common use case, so better for now, while we wait for the spec fix to come in, to prefer the HTTP API.

Relates to #26637.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*