fix(lambda): use of `currentVersion` fails on upgrade
Between version `2.87.0` and version `2.88.0`, the hash calculation
used to make sure that `fn.currentVersion` is automatically updated
when a new version of the Lambda Function is deployed changed.

This causes a creation of a new Version upon upgrading CDK, but that new
Version creation will fail because the underlying Function hasn't
changed.

The change was due to property ordering used in calculating a
unique hash for the Function configuration.

This change restores the property ordering to the pre-2.88.0 behavior.

Fixes #26739.