additional warnings and documentation
Adds to the `externalModules` doc string to tell the default is `[]`
when `NODEJS_LATEST` runtime is used.

Adds to the README to detail that `externalModules` defaults to empty
when using `NODEJS_LATEST` and that `NODEJS_LATEST` is the default when
the feature flag is enabled. Also detail that `NODEJS_LATEST` is
designed to update as new versions are available and packages vended
with the environment may change.

Add a warning that is emitted when using `NODEJS_LATEST` and any
dependencies are external.