fix(eks): undefined constructor in custom resource
When we [upgraded the proxy-agent v6](https://github.com/aws/aws-cdk/pull/26722)
we did not update how we were importing from the package. Where
[in 5.0 commonjs exports were used](https://github.com/TooTallNate/node-proxy-agent/blob/5.0.0/index.js#L25)
, in [6.0 a named esm export is used](https://github.com/TooTallNate/proxy-agents/blob/proxy-agent%406.1.0/packages/proxy-agent/src/index.ts#L71).

Updated to use a named import statement instead of a default require.
Typescript also confirms the old import was an error now since we use
`import` instead of `require`.

Fixes: #26771