fix(ecr): autoDeleteImages fails when repository is renamed (#26742)
This PR fixes the bug that ECRAutoDeleteImages fails on repo rename.

The customResource depends on the role, and when the repository name changes, the role is updated to match the new repository instead of the old one, before customResource runs and the old repository is deleted.

It was difficult to delete the old repo before the role update ran, so I changed the resource of the role to a wildcard.

Closes #26711.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*