fix(redshift): adding distKey to an existing table fails deployment (#26789)
The current implementation executes a `CREATE TABLE` if the table has no `distKey` specified and one is added, or if a `distKey` is present and is removed.
The resulting table is created with the same name, causing the update operation to fail.
This fixes the problem by using [`ALTER TABLE`](https://docs.aws.amazon.com/redshift/latest/dg/r_ALTER_TABLE.html) when adding/removing `distKey` on a table update. 

Closes #26733.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*