fix(custom-resources): bad payload deserialization (#26797)
When `AwsCustomResource` was updated to use sdkv3, it wasn't updated to correctly handle the change in returned value when calling `lambda.invokeFunction`. The payload property was changed from type `Buffer` to `Uint8Array` to increase compatibility in browsers.

Added a check in the `flatten` function used to format payloads in `AwsCustomResource`'s runtime to correctly decode values if they are a typed array or a buffer.

Created a new integ test which fails to deploy if the payload property is not serialized correctly and able to be made a Cfn output.

Fixes: #26730

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*