fix(aws-lambda-python): poetry bundling is broken after Aug 20
The Dockerfile that builds an executable container to run `poetry` in
installs the latest version of `poetry` on a Python 3.7 Docker image.

Unfortunately `poetry 1.6.0`, released August 20, drops support for
Python 3.7 which means that all poetry bundling jobs started failing
on that date.

Upgrade to a base image of Python 3.9 to make the newest poetry work,
and lock the poetry version in place by using a version specifier
so the command is not dependent on the date it is run.