fix(lambda-python): poetry bundling is broken after Aug 20 (#26823)
The Dockerfile that builds an executable container to run `poetry` in installs the latest version of `poetry` on a Python 3.7 Docker image.

Unfortunately `poetry 1.6.0`, released August 20, drops support for Python 3.7 which means that all poetry bundling jobs started failing on that date. `1.6.0` also does not work on Python 3.8, which I don't see mentioned in the change log, but our integration tests that use Python 3.8 are also broken.

Lock the poetry version back to `1.5.1` by using a version specifier so the command is not dependent on the date it is run.

This is the last version that supported both Python 3.7 and 3.8.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*