refactor(lambda): missing Lambda VPC and security validation (#26528)
Specifying the `securityGroups` property requires a `vpc`.

This fix adds validation for the case when a `vpc` is not specified, but `securityGroups` is.
It also adds validation for the case when both `securityGroups` and `allowAllOutbound` are specified (`allowAllOutbound` should be configured in the SGs).

**Question for the reviewers**
How should we handle the case of an empty list in `securityGroups`? (eg `securityGroups: []`)

Closes #26508.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*