feat(core): support bundling for single non-archive file (#26106)
This PR adds a possibility to use local bundling and let the final asset be one single file that is uploaded as-is.

This can be used for several types of assets, e.g. AppSync functions.

In contrast to Lambda, these functions to not expect the asset to be a zip file.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*