feat(cloudfront): Add RealtimeLogConfig to Distribution (#26808)
Pull request to implement RealtimeLogConfig on CloudFront Distribution.
As well as being able to use `RealtimeLogConfigArn` on `Distribution`, I also added `RealtimeLogConfig` as a Construct so that the Cfn is abstracted too.

The reason this is needed so that we can add realtime log config to a `Distribution` without reverting to using `CfnDistribution`.

Closes #<issue number here>.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*