fix(sqs): encryptionType is incorrect when encryptionMasterKey is provided (#26886)
In https://github.com/aws/aws-cdk/issues/19796 we added additional validation to sns subscription.
For that purpose the Queue `encryptionType` was exposed as a public property.
However the PR forgot to take into account that the provided `encryption` property is automatically changed when a `encryptionMasterKey` is provided.

This PR ensures that the public `encryptionType` has the correct value.

Additionally, adds a warning for an incorrect configuration scenario where `encryptionMasterKey` is provided together with an `encryption` other than QueueEncryption.KMS.
This feature was supposed to allow users to simply provide an encryption key and have the encryption type being selected automatically.
However it now unintentionally allows for wrong configurations that are silently fixed, e.g. setting QueueEncryption.UNENCRYPTED and providing an encryption key.
The warning keeps backwards compatibility, but instructs users to fix their configuration.

Closes #26719

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*