docs(lambda-event-sources): remove incorrect description regarding receiveMessageWaitTime (#26882)
Currently, the document for aws_lambda_event_sources module includes the following description.
https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_lambda_event_sources-readme.html#sqs
> receiveMessageWaitTime: Will determine long poll duration. The default value is 20 seconds.

However, from SQS perspective, the default value is 0. So, the above description is incorrect.
https://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_SetQueueAttributes.html
> ReceiveMessageWaitTimeSeconds – The length of time, in seconds, for which a ReceiveMessage action waits for a message to arrive. Valid values: An integer from 0 to 20 (seconds). Default: 0.

Also, when we use SQS queue as Lambda's source, Lambda uses long polling regardless of the queue's ReceiveMessageWaitTimeSeconds setting. 
https://docs.aws.amazon.com/lambda/latest/dg/with-sqs.html#events-sqs-scaling
> For standard queues, Lambda uses long polling to poll a queue until it becomes active.

So, in this context, `receiveMessageWaitTime` prop for Queue construct does not affect the behavior of Lambda EventSource. To avoid confusion, this PR remove the description regarding `receiveMessageWaitTime` from document.

Closes #24795

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*