fix(route53): IHostedZone cannot be used for ses.Identity.publicHostedZone anymore (#26888)
`Identity.publicHostedZone` takes an `IPublicHostedZone`, but because of TypeScript structural typing it would also accept an `IHostedZone`. 

When in [this PR](https://github.com/aws/aws-cdk/pull/26333) the `grantDelegation` method was added to the `IPublicHostedZone` interface, this passing was no longer allowed and code that used to work on accident, no longer works.

For example:
```
const zone = HostedZone.fromHostedZoneId(stack, 'Zone', 'hosted-id');
const sesIdentity = ses.Identity.publicHostedZone(zone);
```

It raises an error because the imported `zone` does not implement the `grantDelegation` method.

This fix moves the `grantDelegation` method declaration into the `IHostedZone` interface and makes it available to all imported zones.

Closes #26872.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*