fix(servicecatalog): only one ProductStack per Portfolio can use assets (#26885)
To move the assets into the target bucket, a `BucketDeployment` was created for every `ProductStack`. Instead, we will now share the same `BucketDeployment` for all `ProductStacks`.

This moves the custom resource around in the construct tree (leading to replacement), but that is a safe operation. Nothing will be destroyed by this move.

Closes #25189.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*