fix(cli): asset not uploaded with different synthesizer configs
If the same asset is used in 2 stacks that use different synthesizer
configurations for publishing (for example, by using a different prefix)
the asset will only be uploaded once instead of twice.

We used to make the assumption that it was okay to use the destination
ID as token of uniqueness. This is true inside a single manifest, but
does not hold when there is more than stack that each have a manifest:
both may have the destination ID `current_account:current_region`, but
have different parameters for each destination.

Instead, we calculate a content hash over the destination definition
itself. That way, if the definitions are different we will create
different nodes for each of them.

Fixes #25927.