chore: fix tests that check package versions in the pipeline (#26934)
They execute correctly locally and in a PR build, but not in the Pipeline because the `framework-integ` package gets an `-alpha.999` suffix (whereas the actual constructs get an `-rc.0` suffix).

The test used to read the expected version from its own `package.json`. It will now read the expected version from `aws-cdk-lib`'s `package.json`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*