fix(custom-resources): empty response causes deployment failure
The payload response of a Lambda used to be a `string`, and could
occasionally be `""`, which we detected and special-case parsed to
an empty object.

In SDKv3, the payload response of a Lambda changed to type `Uint8Array`,
but a `Uint8Array(0)` doesn't check as *falsey*, so we'd decode it to
`""` and then the `JSON.parse()` of that would fail.

First decode, then check the string for emptyness.

Fixes #26429.