fix(route53): `CrossAccountZoneDelegationRecord` is still using Node 16 (#26980)
Updates the `CrossAccountZoneDelegationRecord` construct to use sdk v3 / node 18. 

This is identical to changes in #26212, except for hardcoding a region into the `assumeRole` sdk call. This may not be the ideal solution, but should not break specific configurations.

That specific configuration, as #26593 pointed out, was that the original update was a breaking change if the construct was deployed into an opt-in region, and the parent zone did not have that opt-in region enabled.

This PR removes the semi-hidden `@aws-cdk/aws-route53:useRegionalStsEndpoint` feature flag, as it was based on a confusion on why things used to work. We now pick the correct endpoint manually.

Closes #26976.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*